put MazeSolver into mazegen
